requirements.txt
dash==1.20.0
discum==1.1.0
Django==3.0.5
django_plotly_dash==1.6.6
environs==9.3.2
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
plotly==5.6.0
python-dotenv==0.20.0
toml==0.10.2
celery
django-crispy-forms
django-phonenumber-field
django-bootstrap4
django-celery-results
django-celery-beat
django-ckeditor
djongo
sorl-thumbnail
social-auth-app-django
